fix: modulo de projeto e tarefa p/ store
